Fix 30+ edge cases, logic errors, and bugs across the codebase
Critical gameplay fixes:
- SeekAirGoal: flowing water no longer treated as breathable; guards navigate
  to the air block itself instead of the block below it (still underwater)
- RaidTacticsGoal: offensive guards now call setTarget() so they actually
  attack raiders instead of just walking up and stopping
- GuardEntity: weapon sharpness levels were inverted (level 0 got Sharpness 3,
  level 1 got Sharpness 1); formation type now restored from save data
- GuardArrowEntity: arrows no longer hit the guard's own owner player;
  removed onHit discard that broke piercing enchantment
- GuardBowAttackGoal: targetVisibleTicks no longer decrements unboundedly;
  goal stops properly when target dies
- canAcceptOwnerAlert: inverted logic fixed so low-health idle guards
  respond to owner alerts (previously they refused)

Reputation system fixes:
- Reputation decay now trends toward 0.50 default instead of 0.0
- Hostile threshold lowered to 0.30 (was identical to trust at 0.50,
  meaning any negative delta instantly made guards hostile)
- resetReputation now resets to DEFAULT_REPUTATION instead of 0.0

State management fixes:
- GuardOwnershipIndex: cleanup now removes dead/null guards instead of
  skipping them; added clearAll() for server stop lifecycle
- GuardTacticsState: normalizeRow capped at 256 to prevent OOM from
  malicious row indices
- Static maps cleared on server stop to prevent stale data across restarts
- DEBUG_PATH_HASH_CACHE changed to ConcurrentHashMap for thread safety

Navigation and goal fixes:
- GuardRallyGoal: added distance check to stop re-pathing every tick
- CrowdControlGoal: no longer falls back to guard's own position as anchor
- FormationFollowOwnerGoal: uses guard's world instead of owner's world
  for pathfinding; resets progress sampling on start(); dimension check added
- ElectLeaderGoal: skips first tick (age 0) to avoid premature elections
- GuardNavigation: early null return for null target in findPathTo
- PerimeterPatrolGoal: Math.floorMod for consistent patrol rotation

Entity fixes:
- GuardEntity.setHome(null) no longer throws NPE
- isSameGroup(null) now returns false instead of true
- syncHomeData() called after readCustomData to sync to DataTracker
- addGroup returns -1 on failure instead of valid index 0

Client fixes:
- Debug label colors now include alpha channel (0xFF prefix) so text is
  visible instead of fully transparent
- setGroupColor no longer marks dirty when value unchanged
- saveAssignments clears dirty state even when network handler is null
- isGenericGroupName matches actual defaults (Alpha/Beta/Group N)
- DragHandler scale uses explicit 3-arg form for Z safety
- Scrollbar thumb height clamped to track height
- commitGroupRename defaults to "Alpha" matching server defaults
- VillageManagerHandler.calculateDensity casts to double before multiply
  to prevent int overflow
- GuardShopInventory bulk purchase no longer forces minimum of 1

+++ src/main/java/com/guardvillagers/village/VillageManagerHandler.java
@@ -249,7 +249,7 @@ public final class VillageManagerHandler {
 	}
 
 	private static double calculateDensity(VillageDescriptor village, int villagerCount) {
-		double area = Math.max(1.0D, village.bounds().getBlockCountX() * village.bounds().getBlockCountZ());
+		double area = Math.max(1.0D, (double) village.bounds().getBlockCountX() * village.bounds().getBlockCountZ());
 		return villagerCount / (area / 256.0D);
 	}
